% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Simulation of training sessions and associated performances
Training load and rest between training sessions variables related to it}
\format{
A data frame with 150 rows and 3 variables:
\describe{
\item{training_load}{Session training load, in arbitrary units}
\item{rest}{Time rest between two consecutive training sessions, in days units}
\item{perf}{A performance, in arbitrary units}
\item{datetime}{A datetime object}
}
}
\usage{
example_data
}
\description{
Simulation of training sessions and associated performances
Training load and rest between training sessions variables related to it
}
\keyword{datasets}
