% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init_perf}
\alias{init_perf}
\title{Initial performance}
\usage{
init_perf(data, target)
}
\arguments{
\item{data}{is a data frame object that contains at least training loads, performances and time between two consecutive sessions.}

\item{target}{is a character that indicates the performances column name.}
}
\value{
a numerical value
}
\description{
This function extracts the first performance of a time ordered data frame object.
}
\examples{
init_perf(data = example_data, target = "perf")

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
