% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{real_perf}
\alias{real_perf}
\title{Extract known performances}
\usage{
real_perf(data, target)
}
\arguments{
\item{data}{is a data frame object that contains training loads, performances and time between two consecutive sessions}

\item{target}{is a character vector that indicates the performances column name.}
}
\value{
a vector of numerical values
}
\description{
Extract known performances
}
\examples{
real_perf(data = example_data, target = "perf")

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
