% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RSS}
\alias{RSS}
\title{Residual sum of squares.}
\usage{
RSS(data, theta, target, vars)
}
\arguments{
\item{data}{A data frame object that contains at least training loads, performances and time between two consecutive sessions.}

\item{theta}{A vector of parameters \emph{P0}, \emph{k1}, \emph{k3}, \emph{tau1}, \emph{tau2}, \emph{tau3}.}

\item{target}{A character that indicates the performance column name.}

\item{vars}{A list that contains \emph{input} (i.e. session training loads) and \emph{time} (i.e. the time between two consecutive inputs) numerical vectors.}
}
\value{
a numerical value
}
\description{
This function calculates the residual sum of squares from observations and predicted performances.
}
\examples{
RSS(data = example_data, theta= c(10, 0.1, 0.01, 40, 20, 5), target = "perf", vars = list("input" = example_data$training_load, "time" = example_data$rest))

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
