% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sysmod}
\alias{sysmod}
\title{Dose-response modeling}
\usage{
sysmod(
  data,
  vars,
  target,
  date_ID,
  specify = NULL,
  validation = validation,
  optim.method
)
}
\arguments{
\item{data}{is a data frame object that contains at least training loads, performances and time between two consecutive sessions.}

\item{vars}{is a list that contains \emph{input} (i.e. session training loads) and \emph{time} (i.e. the time between two consecutive inputs) numerical vectors.}

\item{target}{is a character that indicates the performances column name.}

\item{date_ID}{is a character that indicates the date time object name.}

\item{specify}{default is \code{"NULL"} If \code{"NULL"}, a list of numeric vector of initial values for \emph{P0}, \emph{k1}, \emph{k3}, \emph{tau1}, \emph{tau2}, \emph{tau3} parameters,
\emph{lower} and \emph{upper} bounds has to be specified. \emph{PO} is the initial level of performance and can be extracted through the function \link{init_perf}.}

\item{validation}{default is \code{"NULL"}. A list that contains the method used for validation (see details), initial window numerical value, horizon numerical values and
a logical for the fixed window.}

\item{optm.method}{is a character that indicates which method has to be used for parameter optimization (see details).}
}
\value{
A list describing the model output and its performances.
}
\description{

}
\details{
validation can be used to learn model and evaluate within a cross-validation procedure. Methods available are \code{"TS-CV"} for time-series cross-validation.
For \code{"TS-CV"} method, specify numeric values for \code{"initialWindow"}, \code{"horizon"} and logical term for the \code{"fixedWindow"} within a list.
Each of the optimization algorithm with constraints and used by the optimx package can be specified in optim.method.
}
\note{
Model performances (RMSE, MAE and R squared) are calculated on test data for validation = c("simple", "TS-CV").
}
\examples{
sysmod(data = example_data, vars = list("input" = example_data$training_load, "time" = example_data$rest), target = "perf", specify = NULL,
validation = list("initialWindow" = 50, "horizon" = 15, "fixedWindow" = FALSE), optim.method = "nlm", date_ID = "datetime")

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
