% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adaptation_fn}
\alias{adaptation_fn}
\title{Adaptations response modeling}
\usage{
adaptation_fn(data, k1, tau1, vars)
}
\arguments{
\item{data}{A data frame object that contains at least training loads, performances and time between two consecutive sessions.}

\item{k1}{The gain term for adaptation occurrences.}

\item{tau1}{The time constant of the adaptations exponential decrease.}

\item{vars}{A list that contains \emph{input} (i.e. session training loads) and \emph{time} (i.e. the time between two consecutive inputs) numerical vectors.}
}
\value{
a vector of numerical values
}
\description{
To model adaptation induced by each training session.
}
\examples{
adaptation_fn(data = example_data, k1 = 0.5, tau1 = 40, vars = list("input" = example_data$training_load, "time" = example_data$rest))

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
