% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init_perf}
\alias{init_perf}
\title{Extract the first performance}
\usage{
init_perf(data, target)
}
\arguments{
\item{data}{is a data frame object that contains training loads, performances and time between two consecutive sessions}

\item{target}{is a character that indicates the performances column name.}
}
\value{
a numeric value
}
\description{
Extract the first performance
}
\examples{
init_perf(data = example_data, target = "perf")

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
