% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perf_model}
\alias{perf_model}
\title{Function to model the performance following Busso Variable dose-response model (Busso, 2003)}
\usage{
perf_model(data, P0, k1, k3, tau1, tau2, tau3, vars, target)
}
\arguments{
\item{data}{is a data frame object that contains training loads, performances and time between two consecutive sessions}

\item{P0, }{@param k1, @param k2, @param tau1, @param tau2, @param tau3 denote the basic level of performance, gains terms and time constants used in the model}

\item{vars}{denotes a list that contains input and time numeric vectors. Input refers to session training loads and time refers to the time between two consecutive inputs.}

\item{target}{is a character vector that indicates the performances column name.}
}
\value{
a vector of numerical values
}
\description{
Function to model the performance following Busso Variable dose-response model (Busso, 2003)
}
\examples{
perf_model(data = example_data, PO = 10, k1 = 0.1, k3 = 0.01, tau1 = 40, tau2 = 20, tau3 = 5, vars = list("input" = example_data$training_load, "time" = example_data$rest),
target = "perf")

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
