% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{real_perf}
\alias{real_perf}
\title{Extract observations}
\usage{
real_perf(data, target)
}
\arguments{
\item{data}{A data frame object that contains at least training loads, performances and time between two consecutive sessions.}

\item{target}{A character vector that indicates the performances column name.}
}
\value{
a vector of numerical values
}
\description{
This function extracts known performances from a data frame object.
}
\examples{
real_perf(data = example_data, target = "perf")

}
\author{
Frank Imbach \href{mailto:frankimbach@gmail.com}{frankimbach@gmail.com}
}
