% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sysmod}
\alias{sysmod}
\title{The main function that optimises parameters used in the Variable dose-response model (Busso, 2003), predicts values and calculates model performance.}
\usage{
sysmod(
  data,
  vars,
  target,
  date_ID,
  specify = NULL,
  validation = validation,
  optim.method
)
}
\arguments{
\item{data}{is a data frame that contains training loads and performances}

\item{vars}{denotes a list that contains input and time numeric vectors. Input refers to session training loads and time refers to the time between two consecutive inputs.}

\item{target}{is a character vector that indicates the performances column name.}

\item{specify}{default is NULL. If TRUE, a list of numeric vector of initial values for P0, k1, k3, tau1, tau2, tau3 parameters, lower and upper bounds has to be specified.
PO is the initial level of performance and can be extracted through the function init_perf}

\item{validation}{default is NULL. A list that contains the method used for validation (see details), initial window numerical value, horizon numerical values and
a logical for the fixed window.}

\item{optm.method}{is a character that indicates which method has to be used for parameter optimisation (see details)}
}
\description{
The main function that optimises parameters used in the Variable dose-response model (Busso, 2003), predicts values and calculates model performance.
}
\details{
validation can be used to learn model and evaluate within a cross-validation procedure. Methods available are c("TS_CV") for time-series cross-validation.
For "TS-CV" method, specify numeric values for "initialWindow", "horizon" and logical term for the fixed window within a list.
Each of the optimization algorithm with constraints and used by the optimx package can be specified in @param optim.method.
}
